requirements.txt
langchain
langchain-core
langchain-ollama
streamlit
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
beautifulsoup4
pypdf
chromadb
faiss-cpu
wikipedia
arxiv
langchainhub
groq
langchain-groq